+++ sshd-core/src/test/java/org/apache/sshd/common/auth/AuthenticationTest.java
@@ -36,8 +36,10 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
+import java.util.function.Consumer;
 
 import org.apache.sshd.client.SshClient;
+import org.apache.sshd.client.auth.hostbased.HostBasedAuthenticationReporter;
 import org.apache.sshd.client.auth.hostbased.HostKeyIdentityProvider;
 import org.apache.sshd.client.auth.keyboard.UserInteraction;
 import org.apache.sshd.client.auth.password.PasswordAuthenticationReporter;
@@ -72,6 +74,7 @@ import org.apache.sshd.common.util.security.SecurityUtils;
 import org.apache.sshd.core.CoreModuleProperties;
 import org.apache.sshd.server.ServerFactoryManager;
 import org.apache.sshd.server.SshServer;
+import org.apache.sshd.server.auth.hostbased.HostBasedAuthenticator;
 import org.apache.sshd.server.auth.keyboard.DefaultKeyboardInteractiveAuthenticator;
 import org.apache.sshd.server.auth.keyboard.InteractiveChallenge;
 import org.apache.sshd.server.auth.keyboard.KeyboardInteractiveAuthenticator;
@@ -773,17 +776,72 @@ public class AuthenticationTest extends BaseTestSupport {
 
     @Test // see SSHD-620
     public void testHostBasedAuthentication() throws Exception {
-        String hostClienUser = getClass().getSimpleName();
+        AtomicInteger invocationCount = new AtomicInteger(0);
+        testHostBasedAuthentication(
+                (
+                        session, username, clientHostKey, clientHostName, clientUsername,
+                        certificates) -> invocationCount.incrementAndGet() > 0,
+                session -> {
+                    /* ignored */ });
+        assertEquals("Mismatched authenticator invocation count", 1, invocationCount.get());
+    }
+
+    @Test   // see SSHD-1114
+    public void testHostBasedAuthenticationReporter() throws Exception {
+        AtomicReference<String> hostnameClientHolder = new AtomicReference<>();
+        AtomicReference<String> usernameClientHolder = new AtomicReference<>();
+        AtomicReference<PublicKey> keyClientHolder = new AtomicReference<>();
+        HostBasedAuthenticator authenticator
+                = (session, username, clientHostKey, clientHostName, clientUsername, certificates) -> {
+                    return Objects.equals(clientHostName, hostnameClientHolder.get())
+                            && Objects.equals(clientUsername, usernameClientHolder.get())
+                            && KeyUtils.compareKeys(clientHostKey, keyClientHolder.get());
+                };
+
+        HostBasedAuthenticationReporter reporter = new HostBasedAuthenticationReporter() {
+            @Override
+            public void signalAuthenticationAttempt(
+                    ClientSession session, String service, KeyPair identity, String hostname, String username, byte[] signature)
+                    throws Exception {
+                hostnameClientHolder.set(hostname);
+                usernameClientHolder.set(username);
+                keyClientHolder.set(identity.getPublic());
+            }
+
+            @Override
+            public void signalAuthenticationSuccess(
+                    ClientSession session, String service, KeyPair identity, String hostname, String username)
+                    throws Exception {
+                assertEquals("Host", hostname, hostnameClientHolder.get());
+                assertEquals("User", username, usernameClientHolder.get());
+                assertKeyEquals("Identity", identity.getPublic(), keyClientHolder.get());
+            }
+
+            @Override
+            public void signalAuthenticationFailure(
+                    ClientSession session, String service, KeyPair identity,
+                    String hostname, String username, boolean partial, List<String> serverMethods)
+                    throws Exception {
+                fail("Unexpected failure signalled");
+            }
+        };
+
+        testHostBasedAuthentication(authenticator, session -> session.setHostBasedAuthenticationReporter(reporter));
+    }
+
+    private void testHostBasedAuthentication(
+            HostBasedAuthenticator delegate, Consumer<? super ClientSession> preAuthInitializer)
+            throws Exception {
+        String hostClientUser = getClass().getSimpleName();
         String hostClientName = SshdSocketAddress.toAddressString(SshdSocketAddress.getFirstExternalNetwork4Address());
         KeyPair hostClientKey = CommonTestSupportUtils.generateKeyPair(
                 CommonTestSupportUtils.DEFAULT_TEST_HOST_KEY_PROVIDER_ALGORITHM,
                 CommonTestSupportUtils.DEFAULT_TEST_HOST_KEY_SIZE);
-        AtomicInteger invocationCount = new AtomicInteger(0);
         sshd.setHostBasedAuthenticator((session, username, clientHostKey, clientHostName, clientUsername, certificates) -> {
-            invocationCount.incrementAndGet();
-            return hostClienUser.equals(clientUsername)
+            return hostClientUser.equals(clientUsername)
                     && hostClientName.equals(clientHostName)
-                    && KeyUtils.compareKeys(hostClientKey.getPublic(), clientHostKey);
+                    && KeyUtils.compareKeys(hostClientKey.getPublic(), clientHostKey)
+                    && delegate.authenticate(session, username, clientHostKey, clientHostName, clientUsername, certificates);
         });
         sshd.setPasswordAuthenticator(RejectAllPasswordAuthenticator.INSTANCE);
         sshd.setKeyboardInteractiveAuthenticator(KeyboardInteractiveAuthenticator.NONE);
@@ -796,16 +854,16 @@ public class AuthenticationTest extends BaseTestSupport {
             org.apache.sshd.client.auth.hostbased.UserAuthHostBasedFactory factory
                     = new org.apache.sshd.client.auth.hostbased.UserAuthHostBasedFactory();
             // TODO factory.setClientHostname(CLIENT_HOSTNAME);
-            factory.setClientUsername(hostClienUser);
+            factory.setClientUsername(hostClientUser);
             factory.setClientHostKeys(HostKeyIdentityProvider.wrap(hostClientKey));
 
             client.setUserAuthFactories(Collections.singletonList(factory));
             client.start();
-            try (ClientSession s = client.connect(getCurrentTestName(), TEST_LOCALHOST, port)
+            try (ClientSession session = client.connect(getCurrentTestName(), TEST_LOCALHOST, port)
                     .verify(CONNECT_TIMEOUT)
                     .getSession()) {
-                s.auth().verify(AUTH_TIMEOUT);
-                assertEquals("Mismatched authenticator invocation count", 1, invocationCount.get());
+                preAuthInitializer.accept(session);
+                session.auth().verify(AUTH_TIMEOUT);
             } finally {
                 client.stop();
             }